requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
django-ajax-datatable==4.4.5
django-cors-headers==3.14.0
django-cron==0.6.0
django-crontab==0.7.1
django-model-utils==4.3.1
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
lxml==4.9.2
numpy==1.24.2
pandas==1.5.3
psycopg2-binary==2.9.5
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
urllib3==1.26.15
